Move unification into its own module
